Moved open and close buttons in blind preferences to separate buttons
