+++ app/src/main/AndroidManifest.xml
@@ -14,8 +14,6 @@
         android:theme="@style/Theme.HelioApp"
         android:usesCleartextTraffic="true"
         tools:ignore="AllowBackup,UnusedAttribute">
-        <activity android:name=".AddScheduleCloseActivity" />
-        <activity android:name=".AddScheduleOpenActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
