+++ app/src/main/java/com/helio/app/MainActivity.java
@@ -2,10 +2,13 @@ package com.helio.app;
 
 import android.os.Bundle;
 import android.os.Handler;
+import android.view.Menu;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
+import androidx.navigation.fragment.NavHostFragment;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
@@ -22,23 +25,19 @@ import java.util.Map;
 
 public class MainActivity extends AppCompatActivity {
     private Map<Integer, Motor> motors;
+    AppBarConfiguration appBarConfiguration;
 
+    @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        BottomNavigationView navView = findViewById(R.id.nav_view);
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_schedule, R.id.blinds_settings)
-                .build();
-        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
 
-        // This makes the bar at the top change when you use the navigation (changes the title text)
+        NavHostFragment navHostFragment =
+                (NavHostFragment) getSupportFragmentManager().findFragmentById(R.id.nav_host_fragment);
+        assert navHostFragment != null;
+        NavController navController = navHostFragment.getNavController();
+        appBarConfiguration = new AppBarConfiguration.Builder(navController.getGraph()).build();
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
-
-        // This makes the fragment change when you press the navigation buttons
-        NavigationUI.setupWithNavController(navView, navController);
 
         fetchState();
     }
