+++ app/src/main/java/com/helio/app/ui/blind/SingleBlindSettingsFragment.java
@@ -1,0 +1,23 @@
+package com.helio.app.ui.blind;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
+import com.helio.app.R;
+
+public class SingleBlindSettingsFragment extends Fragment {
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        View returnView = inflater.inflate(R.layout.fragment_single_blind_settings, container, false);
+
+        return returnView;
+    }
+
+}
