+++ app/src/main/java/com/helio/app/ui/blinds/BlindsRecViewAdapter.java
@@ -1,4 +1,4 @@
-package com.helio.app.ui.blindSettings;
+package com.helio.app.ui.blinds;
 
 import android.content.Context;
 import android.view.LayoutInflater;
