--- app/src/main/java/com/helio/app/ui/controlling/ControlViewModel.java
@@ -1,7 +1,0 @@
-package com.helio.app.ui.controlling;
-
-import androidx.lifecycle.ViewModel;
-
-public class ControlViewModel extends ViewModel {
-    // TODO: Implement the ViewModel
-}
