--- app/src/main/res/layout/activity_add_schedule_open.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <!--Title-->
-    <TextView
-        android:id="@+id/tv_current_setting"
-        android:text="Set opening time"
-        android:textSize="40sp"
-        android:layout_marginStart="40dp"
-        android:textColor="@color/black"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-
-
-    <!--Set-->
-    <Button
-        android:id="@+id/btn_submit"
-        android:layout_width="300dp"
-        android:layout_height="100dp"
-        android:text="set"
-        android:textSize="20sp"
-        android:layout_below="@id/tv_current_setting"
-        android:layout_marginTop="350dp"
-        android:layout_marginStart="50dp"/>
-
-    <!--Back-->
-    <Button
-        android:id="@+id/btn_back"
-        android:layout_width="300dp"
-        android:layout_height="100dp"
-        android:text="back"
-        android:textSize="20sp"
-        android:layout_below="@id/tv_current_setting"
-        android:layout_marginTop="450dp"
-        android:layout_marginStart="50dp"/>
-
-    <!--Selecting time-->
-    <TimePicker
-        android:id="@+id/tp_time1"
-        android:timePickerMode="spinner"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/tv_current_setting"
-        android:layout_marginStart="70dp"
-        android:layout_marginTop="80dp"
-        android:background="@drawable/block"
-        />
-
-</RelativeLayout>
