+++ app/src/main/res/layout/activity_main.xml
@@ -6,17 +6,13 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/nav_view"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="0dp"
-        android:layout_marginEnd="0dp"
-        android:background="?android:attr/windowBackground"
-        app:layout_constraintBottom_toBottomOf="parent"
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
+        android:layout_height="40dp"
+        android:layout_width="match_parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:menu="@menu/bottom_nav_menu" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <!-- Studio suggests fragment is replaced with FragmentContainerView, but that makes it crash,
     so ignore the warning with tools:ignore. Set the width+height to 0dp so that it fits to size properly. -->
@@ -26,13 +22,11 @@
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:defaultNavHost="true"
-        app:layout_constraintBottom_toTopOf="@id/nav_view"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:navGraph="@navigation/mobile_navigation"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:navGraph="@navigation/main_navigation"
         tools:ignore="FragmentTagUsage" />
-
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
