--- app/src/main/res/layout/control_fragment.xml
@@ -1,136 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.controlling.ControlFragment">
-
-    <!--text-->
-    <TextView
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:text="IP:"
-        android:textColor="@color/black"
-        android:textSize="20dp"
-        android:layout_marginTop="61dp"
-        android:layout_marginLeft="70dp"
-        />
-    <TextView
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:text="Name:"
-        android:textColor="@color/black"
-        android:textSize="20dp"
-        android:layout_marginTop="130dp"
-        android:layout_marginLeft="70dp"
-        />
-    <TextView
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:text="State:"
-        android:textColor="@color/black"
-        android:textSize="20dp"
-        android:layout_marginTop="200dp"
-        android:layout_marginLeft="70dp"
-        />
-
-    <!--IP text box-->
-    <EditText
-        android:id="@+id/et_IPTextBox"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:background="@drawable/block"
-        android:layout_marginTop="50dp"
-        android:layout_marginLeft="60dp"
-        android:inputType="number"
-        android:textColor="@color/black"
-        android:paddingLeft="35dp"
-        />
-
-    <!--name text box-->
-    <EditText
-        android:id="@+id/et_nameTextBox"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:background="@drawable/block"
-        android:layout_marginTop="120dp"
-        android:layout_marginLeft="60dp"
-        android:textColor="@color/black"
-        android:paddingLeft="70dp"
-        />
-
-    <!--current state of blind-->
-    <EditText
-        android:id="@+id/et_currentState"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:background="@drawable/block"
-        android:layout_marginTop="190dp"
-        android:layout_marginLeft="60dp"
-        android:textColor="@color/black"
-        android:inputType="number"
-        android:paddingLeft="65dp"
-        />
-
-    <!--change icon-->
-    <Button
-        android:id="@+id/btn_changeIcon"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="260dp"
-        android:layout_marginLeft="60dp"
-        android:text="change icon"
-        />
-
-    <!--calibration button-->
-    <Button
-        android:id="@+id/btn_calibration"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="330dp"
-        android:layout_marginLeft="60dp"
-        android:text="calibration"
-        />
-
-    <!--sensors-->
-    <Button
-        android:id="@+id/btn_sensors"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="400dp"
-        android:layout_marginLeft="60dp"
-        android:text="sensors"
-        />
-
-    <!--create schedule-->
-    <Button
-        android:id="@+id/btn_createSchdule"
-        android:layout_width="270dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="470dp"
-        android:layout_marginLeft="60dp"
-        android:text="create schedule"
-        />
-
-    <!--save-->
-    <Button
-        android:id="@+id/btn_save"
-        android:layout_width="130dp"
-        android:layout_height="80dp"
-        android:layout_marginTop="530dp"
-        android:layout_marginLeft="60dp"
-        android:text="save"
-        />
-
-    <!--cancel-->
-    <Button
-        android:id="@+id/btn_cancel"
-        android:layout_width="130dp"
-        android:layout_height="80dp"
-        android:layout_marginTop="530dp"
-        android:layout_marginLeft="200dp"
-        android:text="cancel"
-        />
-
-
-</RelativeLayout>
