+++ app/src/main/res/layout/fragment_home.xml
@@ -1,45 +1,45 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    >
+    tools:context=".MainActivity">
 
     <RelativeLayout
+        android:id="@+id/controlAllRelLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/controlAllRelLayout"
-        android:layout_marginHorizontal="20dp"
-        android:layout_centerInParent="true">
+        android:layout_centerInParent="true"
+        android:layout_marginHorizontal="20dp">
 
         <TextView
             android:id="@+id/textControlBlinds"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Control all of your blinds"
-            android:textSize="30sp"
-            android:textStyle="bold"
             android:layout_centerHorizontal="true"
             android:layout_marginBottom="10dp"
-            />
+            android:text="Control all of your blinds"
+            android:textSize="30sp"
+            android:textStyle="bold" />
 
         <Button
             android:id="@+id/buttonRaise"
-            android:layout_below="@+id/textControlBlinds"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:paddingVertical="20dp"
+            android:layout_below="@+id/textControlBlinds"
             android:layout_centerHorizontal="true"
+            android:paddingVertical="20dp"
             android:text="Raise"
             android:textSize="30sp" />
 
         <Button
             android:id="@+id/buttonLower"
-            android:layout_below="@+id/buttonRaise"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:paddingVertical="20dp"
+            android:layout_below="@+id/buttonRaise"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="10dp"
+            android:paddingVertical="20dp"
             android:text="Lower"
             android:textSize="30sp" />
     </RelativeLayout>
