--- app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
-
-    <item
-        android:id="@+id/navigation_schedule"
-        android:icon="@drawable/ic_baseline_access_time_24"
-        android:title="@string/title_schedule" />
-
-    <item
-        android:id="@+id/blinds_settings"
-        android:icon="@drawable/ic_baseline_wifi_24"
-        android:title="@string/title_blinds_settings" />
-
-</menu>
