+++ app/src/main/res/navigation/main_navigation.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation"
+    app:startDestination="@id/homeFragment">
+
+    <fragment
+        android:id="@+id/blindsSettingsFragment"
+        android:name="com.helio.app.ui.blinds.BlindsSettingsFragment"
+        android:label="@string/title_blinds_settings"
+        tools:layout="@layout/fragment_blinds_settings">
+        <action
+            android:id="@+id/action_blindsSettingsFragment_to_singleBlindSettingsFragment"
+            app:destination="@id/singleBlindSettingsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/hubFragment"
+        android:name="com.helio.app.ui.hub.HubFragment"
+        android:label="@string/title_connect_hub"
+        tools:layout="@layout/fragment_hub" />
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.helio.app.ui.blind.SingleBlindSettingsFragment"
+        android:label="@string/title_home"
+        tools:layout="@layout/fragment_home">
+        <action
+            android:id="@+id/action_homeFragment_to_blindsSettingsFragment2"
+            app:destination="@id/blindsSettingsFragment" />
+        <action
+            android:id="@+id/action_homeFragment_to_hubFragment"
+            app:destination="@id/hubFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/singleBlindSettingsFragment"
+        android:name="com.helio.app.ui.blind.SingleBlindSettingsFragment"
+        android:label="@string/title_edit_blind"
+        tools:layout="@layout/fragment_single_blind_settings"/>
+</navigation>
